*High fives all around*
We did it guys!  This version of the game has 0 known bugs and issues with it. There is one more scalability feature I would like to add before I call this the MVP of RGBA (that being scaleable start and end positions relative to the center of the screen). However, I think 0 known bugs and a fully functional app is something pretty crazy and I am extremely happy that I've gotten this far with RGBA.🤘