Have you ever just forgotten to commit your code? Yeah...
So the last thing I worked on in the commit is first time loading and user experience (demo/instructions). This commit also includes some testing code and a addition to the standards for scenes to be able to control the background/overflow scene design. I want to add more levels and more sounds/music before fully releasing.